fix: split blob and OCI policy commands
